+++ arch/arm/src/stm32/stm32_tickless.c
@@ -686,8 +686,8 @@ int up_timer_gettime(struct timespec *ts)
   ts->tv_sec  = sec;
   ts->tv_nsec = (usec - (sec * USEC_PER_SEC)) * NSEC_PER_USEC;
 
-  tmrinfo("usec=%llu ts=(%lu, %lu)\n",
-          usec, (unsigned long)ts->tv_sec, (unsigned long)ts->tv_nsec);
+  tmrinfo("usec=%llu ts=(%jd, %ld)\n",
+          usec, (intmax_t)ts->tv_sec, ts->tv_nsec);
 
   return OK;
 }
@@ -872,8 +872,8 @@ int up_timer_cancel(struct timespec *ts)
       ts->tv_sec  = sec;
       ts->tv_nsec = nsec;
 
-      tmrinfo("remaining (%lu, %lu)\n",
-             (unsigned long)ts->tv_sec, (unsigned long)ts->tv_nsec);
+      tmrinfo("remaining (%jd, %ld)\n",
+              (intmax_t)ts->tv_sec, ts->tv_nsec);
     }
 
   return OK;
@@ -911,8 +911,8 @@ int up_timer_start(const struct timespec *ts)
   uint32_t count;
   irqstate_t flags;
 
-  tmrinfo("ts=(%lu, %lu)\n",
-          (unsigned long)ts->tv_sec, (unsigned long)ts->tv_nsec);
+  tmrinfo("ts=(%jd, %ld)\n",
+          (intmax_t)ts->tv_sec, ts->tv_nsec);
   DEBUGASSERT(ts);
   DEBUGASSERT(g_tickless.tch);
