+++ arch/xtensa/src/esp32s3/esp32s3_oneshot.c
@@ -262,9 +262,8 @@ int esp32s3_oneshot_start(struct esp32s3_oneshot_s *oneshot,
   uint64_t timeout_us;
   int ret = OK;
 
-  tmrinfo("handler=%p arg=%p, ts=(%lu, %lu)\n",
-          handler, arg, (unsigned long)ts->tv_sec,
-          (unsigned long)ts->tv_nsec);
+  tmrinfo("handler=%p arg=%p, ts=(%jd, %ld)\n",
+          handler, arg, (intmax_t)ts->tv_sec, ts->tv_nsec);
 
   DEBUGASSERT(oneshot != NULL);
   DEBUGASSERT(handler != NULL);
